+++ CHANGELOG.md
@@ -31,6 +31,10 @@ If you like this project and find it useful, please consider giving it a star on
 
 ## [3.0.4] - Dev branch
 
+### Breaking changes
+
+- [matterbridge]: Require matterbridge v.3.10.8 with matter v.1.6.0.
+
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>
 
 ## [3.0.3] - 2026-08-30